+++ src/lib/overlay/index.ts
@@ -5,3 +5,5 @@ export { OverlayPosition } from './overlay-position';
 export { OverlayRef } from './overlay-ref';
 export { OverlayService } from './overlay-service';
 export type { VerticalConnectionPos } from './vertical-connection-pos';
+export * from './connected-positions';
+export { registerOverlayKeydown } from './overlay-keyboard-dispatcher';